pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "progressive-diffusion-distillation-with-adaptive-step-curriculum"
version = "0.1.0"
description = "Distills large diffusion models into faster few-step generators using progressive knowledge distillation with adaptive curriculum learning"
authors = [
    {name = "Alireza Shojaei"}
]
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
keywords = [
    "diffusion models",
    "knowledge distillation",
    "curriculum learning",
    "text-to-image",
    "generative ai",
    "deep learning",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "diffusers>=0.25.0",
    "transformers>=4.35.0",
    "accelerate>=0.25.0",
    "peft>=0.7.0",
    "datasets>=2.16.0",
    "Pillow>=10.0.0",
    "scikit-learn>=1.3.0",
    "matplotlib>=3.7.0",
    "PyYAML>=6.0",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
]
mlflow = [
    "mlflow>=2.9.0",
]

[project.urls]
Homepage = "https://github.com/alirezashojaei/progressive-diffusion-distillation-with-adaptive-step-curriculum"

[tool.setuptools]
packages = ["progressive_diffusion_distillation_with_adaptive_step_curriculum"]
package-dir = {"" = "src"}

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--verbose --cov=src --cov-report=term-missing --cov-report=html"

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
